Set default per_page in resources_controller.rb
Prior to this change we used geared pagination default ratios of [15,30,
50,100] for paging, which is suitable for loading data during
infinite scroll but not so much with tabular data, where it is more
reasonable to show fixed amount of rows on each page.
Here we allow children controllers to override per_page in case they
still want to use ratios or define their own fixed `per_page` amounts.