Allow optional passing of options to calculators
For the upcoming "strikethrough prices"/"promotion previews" feature, we
need to allow passing options to calculator's `compute` methods.

For strikethrough prices, we will need the current order and the desired
quantity to be passed into the calculator, and this commit allows
passing any desired options.

Since calculators are polymorphic in all directions and can be used in
all kinds of context, we don't specify which options are possible.

This also does the groundwork for allowing the
SolidusPromotions::Benefit to take options and pass them on to its
calculator.

Because these methods are used a lot in Solidus, adds YARD-style
documentation to the modified methods.