Promotion Configuration: Add #price_conditions
For the upcoming `AdjustPrices` Benefit, we need to have a set of
conditions that can be used with both prices and line items. It's
currently set to be empty, but will be filled as we expand the list of
conditions that can be used for both line items and prices.

Both need to be supported so that promotions don't have surprising
results.