Allow FlatRate discount to be previewed
We need to make sure that the FlatRate discount is not applied twice to
the same line item.

For now, I'm only considering the standard case of the same variant
always being on the same line item here. If the store implementing this
has more complex requirements, please adapt the calculator or write a
custom one.