Add Promotion#previous_lanes
Sometimes, we need to calculate the adjustments given before a
particular promotion runs: When calculating a price, we need a line
item's discountable amount but only up to the previous lane. In order to
get that, we need to know which lanes come before the current promotion.
That's what this commit gives us.