Add `Spree::ZERO` constant
Ruby's `BigDecimal` library does not come with a constant that denotes
zero. We don't need to be creating a new zero every time we need one.
Also, there's even a Rubocop that tells people to use an integer for
integer-type BigDecimal instantiations, but a string for float-type
BigDecimals (See
https://github.com/rubocop/rubocop-performance/blob/master/lib/rubocop/cop/performance/big_decimal_with_numeric_argument.rb)

This all to say: Let's have a constant in the codebase that is zero. We
can just reference it, we don't need so many zeros.