Remove optionality of Spree::RoleUser foreign keys
This is a join table, neither of the keys is optional.

This also adds a foreign key constraint between the role_id column of the
spree_roles_users table and the spree_roles table.

We're not adding one for the users table, because the user class is
configurable and we do not know which table to constrain to.