Make OrderRecalculatorPatch prepend explicit
Prepending a configurable class runs into all kinds of trouble with load
order. If the patch is loaded before the configuration changes, the
newly configured class won't have the patch applied.

This shows up fairly often in tests, but could also be an issue in other
environments if the config is not set to run before all of the code is
eager loaded.

Co-authored-by: Jared Norman <jared@super.gd>
Co-authored-by: Noah Silvera <noah@super.gd>