Switch to attribute_assigns for item totals
Instead of calling update_columns, we should just use assign_attributes
everywhere and rely on our order persistance bubbling down to the line
item when appropriate.

This avoids issues with the in-memory order updater when attempting to
update values on newly created promotional items that won't get
persisted until later in the recalculation flow.

Co-authored-by: Jared Norman <jared@super.gd>
Co-authored-by: Noah Silvera <noah@super.gd>