Add and use Shipment#recalculate_state
We don't want our in-memory order updater to have to persist shipment
state. We also discovered that a portion of the existing
`Shipment#update_state` method has conditionals that could never be hit.

So we're adding a new method to help with the recalculation of the
shipment state because we don't want to modify any of the existing
methods because they're public.

Co-authored-by: Jared Norman <jared@super.gd>
Co-authored-by: Noah Silvera <noah@super.gd>