Create state changes in background job
In many state machines (order, payment and shipment)
we record state changes in the `spree_state_changes` table
as well as in the `OrderUpdater`.

This happens in the same transaction as the model change
itself. In the latest `state_machines` gem version (`> 0.30`)
these record creations cause validation errors and an
infinite validation loop.

Since the state changes are for debugging / analyse purposes
only and are not critical to the actual process of buying
we can easily background them.

In order to keep the correct timeline of events we make sure
that the time is stored on the event time and not background
job runtime.