Add foreign key constraint point to country table
This adds database-level foreign key constraints to tables that
reference the countries table.

If trying to delete a country, and there are addresses referencing that
country, we want to restrict that. Addresses are immutable and must stay
valid.

If there are prices referencing the country, the same applies, we want
to restrict.

If the only thing holding us back from deleting a country is states, we
can delete the state records as well.