FK Migrations: Match on Exception names
When migrating the database in a Solidus context, we usually only have
the database adapter loaded that is being used. Rescuing exceptions from
other adapters will therefore fail with an "uninitialized constant
error". In order to safeguard from this, this commit changes the
matching on error class to matching on their name.