SolidusAdmin: Configure generator namespace early in generator
A previous solution for this bug tried to set the generator namespace in
an app initializer, but that required `rails/generators/base` to be
required - which it is not always when starting up e.g. a server.

I've tried to dig a bit into how this `mattr_accessor :namespace` on
`Rails::Generators::Base` is supposed to be used. It seems like it is
run when loading an engine's generators here[1], but that does not work
in this context.

I don't fully understand how Rails::Generators.namespace is supposed to
be set, but this at least sets it only when needed. There's probably
some magic that we could invoke, but I can't find it documented
anywhere.

[1] https://github.com/rails/rails/blob/main/railties/lib/rails/command/actions.rb#L38