Use `#recalculate_state` in `Shipment#update_state`
While `#determine_state` is technically a public method, it is only used
within the shipment model. We also checked all repos on the `solidusio`
and `solidusio-contrib` GitHub organizations and only found it used
once. So in this commit we remove the method entirely and replace its
single usage with the more useful `Shipment#recalculate_state` method.

Using `#recalculate_state` is beneficial because it assigns the new
state to the shipment in memory. This gives developers more flexibility
to actually do something with the new state information. For example,
when simulating order updates without persisting them.

In some other upcoming work on an in-memory order updater (#5872) we use
the new `Shipment#recalculate_state` method as a public method, within
the new order updater, which we why we think this is a good candidate
for a public method.

Co-authored-by: Chris Todorov <chris@super.gd>