Expose issue with `OrderUpdater#log_state_change`
When the new asynchronous state change logging job was introduced in
\#6304, the attributes saved to `spree_state_change` records was
inadvertently changed to exclude the name of the model whose state was
being changed.

For example, when the order updater calls `log_state_change("payment")`,
the state change record would have the name "payment" on it. Since the
job was introduced all of the state change records created by the order
updater would have a name "order".

This commit simply exposes the change in behaviour so we can understand
and resolve the issue more clearly.

Co-Authored-By: Adam Mueller <adam@super.gd>