Fix issue with splitting backordered units
When splitting backordered inventory units, we were incorrectly sending
all backordered units to the desired shipment, regardless of the
quantity chosen.

Our original work on this changed from computing on_hand units first to
the backordered ones. Unfortunately, we missed that we should only ever
move as many as was requested, not all of the possible backordered ones.