Promotions: First Order condition must not fail on canceled orders
The store I work in lets customers cancel their orders, and create a new
one in its stead. The feature is described as "Edit your order" to
customers. When they do this, the order they then place is perceived to
be their first order, and I think we should respect that.

Also changes the spec to work with actual records rather than mocking
chains.