PercentWithCap calculator: Add YARD docs, do not override #compute
We want `compute` to stay intact, and we want the calculator to document
what kind of items it can discount.

That's why we need to stop inheriting from the `Percent` calculator,
because that will work with shipments and shipping rates, while anything
to do with Distributed Amounts only works on line items.