Tiered Calculators: Transform tier hash in method
Prior to this, transforming the tier hash was done in an anonymous
block, converted everything to a string twice, and was generally not
very legible. I'm aware the new code iterates over the tier list twice,
not once as before, but I think it's a lot more legible this way.