Conditions: Change API
Prior to this, we had a bit of cumbersome way of definining a condition.
We first had to define whether the condition is applicable by
implementing a method, and then we had to define the actual,
polymorphic, eligibility check.

This commit changes the matter such that a condition can simply
implement

```rb
def order_eligible? = order.foo?
```

and the promotion system knows that this condition is applicable for
order checks. Neat!

This - along with the deprecation of `Condition#level` - allows us to
also deprecate the `OrderLevelCondition`, `LineItemLevelCondition`, and
`ShipmentLevelCondition` modules that do nothing but a type check.

Next commit will have the deprecation warnings.

For the upcoming price promotions (strikethrough prices) we need to be
able to test whether a price is eligible for a promotion given a
current order and a quantity. This adds a test for that API.