Deprecate `Level` concerns
The `LineItemLevelCondition`, `OrderLevelCondition`, and
`ShipmentLevelCondition` concerns are awkwardly named, and there's no
adequate thing for conditions that target shipments and shipping rates
with this system (let alone something for prices). With the previous
commit in place and #6357, we can fully deprecate these modules for
Solidus 5.

What this does is:
If a condition includes any of the modules and defines the `eligible?`
method, we emit a deprecation warning telling the implementer to stop
including the module and rename their method. That way the `respond_to?`
logic from the base class kicks in, and we can remove the awkward API.