Extract OrderProduct Condition
This condition is like the `Product` condition with the
`apply_to_line_items` preference set to false.

This condition is true for an order that matches the condition, but does
not restrict the matching line items. This is useful for things like "10
% off lids if you also bought the matching jar".

The intent here is to be able to remove the
`LineItemApplicableOrderCondition`, because that's a hard one to parse
for users.