Extract OrderTaxon condition
We want users to choose whether they want to match
a) orders that contain a taxon
b) orders that contain a taxon AND line items that contain a taxon
c) line items that contain a taxon.

This new class adds case a).

This also removes an unreachable condition and the unrealistic spec for
it.