Add Promotion.lanes_before(lane) and .lanes_before_current_lane
This will return the lanes that came before the current lane. Useful
for finding the discountable amount by lane.

If no lane is present, we just return all lanes, as the expectation is
that any adjustments that are present and promotion adjustments would be
relevant as we're outside the promotion calculation process.