Add amount limit to promotion model
We frequently have a need to construct a promotion that has a cap on how
much discount people can get. While this can be accomplished using
calculators, it's preferable to have this set at the promotion level -
because a promotion can have multiple benefits, and therefore multiple
calculators, storing the limit on the calculators would require
promotion admins to enter the same number multiple times. It's easier to
do this at the promotion level.