Store discounts for current lane on discountable item
If we want to give discounts for the currenly discountable item only up
to a certain limit, we need the item to know what's currently being
calculated. So rather than having variables local to the `DiscountOrder`
service object, we add them to the discountable item as
`current_lane_discounts`, and clear them after having calculated each
lane.