Rename reset_current_discounts to reset_discounts
When we call order.reset_discounts, we don't want to deal with those
that are currently valid/eligible during promotion recalulcation, we
want all discounts to be zero.

We could destroy them, but zeroing them is more efficient, because it's
likely that an adjustment might not change during adjustment
recalculation, in which case we don't need to destroy and re-create.