Change Benefit#discount to dynamically call methods on benefits
Similar to the work in #6360, this improves the DSL for creating
benefits. Rather than having to define one method that cares for
applicability to a discountable, we now simply define that a benefit
`#can_discount?` an object if its public interface has
`#discount_{discountable_type}` defined.

This does not change the implementation of the different methods, but
that will come when we refactor the promotion system to use a single
system of record for discounts. Then, different objects will need
different discount records (line items and shipment use adjustments, but
shipping rates use shipping rate discounts etc.).

I'm not doing the work for adding deprecation warnings here, as this
code was never meant to be overridden up to now.

Also fixes some AI slop in the docs.