Change promotion system to use adjustments
We want the individual benefits to issue exactly the kind of object that
will discount the adjustable object, so we need to generate the right kind
of discount object for each adjustable type.

This allows us to remove the `PersistDiscountedOrder` class, and it
makes the promotion system more understandable and more in line with the
rest of Solidus. Specifically, we can now check for adjustments when
calculating promotion discounts (this is pretty big!), and
`promotion.discounted_amount` can be called from outside the promotion
calculation loop.