Remove inverse_of from Spree::Adjustment#order
This `inverse_of` gives rise to spurious adjustments appearing in the
database if there are newly built and directly marked for destruction.

Since we're never calling `adjustment.order` anywhere in production
code, I am pretty sure we can remove this inverse_of safely.

I believe what happens is that the adjustment gets added to
`order.all_adjustments` on initialization, and marking it for
destruction on the shipment does not carry on to that association as it
does not have the `autosave` property.

Alternatively, we could add `autosave` to `order.all_adjustments`, but
that could surface even weirder bugs.