Add Promotion.within_lane
This allows setting the current promotion lane to the current thread,
and run computations with the current promotion lane set.

This is useful for calculating what the discounted amount of a
promotable is within e.g. the `default` lane even if there are
adjustments for the `post` lane.

I'm using `Thread.current` here because we don't always have an order to
store the current lane on.