Fix Calculator Default Fields
For all calculators except the `PercentWithCap` one we have a custom
`calculator_fields` partial. PercentWithCap does not need that - it can
fall back to the default calculator fields partial. However, that
partial does not render currently, because `name` is a symbol that does
not respond to `humanize`.

Adds a system spec as well.