Taxon condition: Allow non-persisted line items
This refactors the taxon conditions to rely on in-memory properties of
line items rather than on the current database state of the order. The
database state of the conditions is still necessary - we're not
changing conditions while calculating promotions.

The change to the order condition will result in a lot of queries to
line item product classifications - but since it's an order level
condition, that is only going to happen once. For line item level
condition, we're actually removing an n+1 here, because we're
instacaching the list of taxons and descendants on the condition
instance.