Hydrate taxon_ids_with_children cache in taxon condition
The taxon condition instantiates new in-memory conditions for checking
at the order level and at the line item level. The array of arrays of
taxons ids with their respective descendants is instacached on the
instance. In order to prevent loading this array n times (once for each
line item, and once for the order), we hydrate the instance caches on
the in-memory conditions.