Create OrderAdjuster::NullifyOrderDiscounts
This service object sets all discounts on an order - adjustments on
shipments and line items, and shipping rate discounts on shipping rates
- to an amount of zero. This service class is intended to run within the
order adjuster, before promotion calculation. Promotion calculation will
then re-calculate the amounts, with the effect that adjustments that
need to change will be changed, while adjustments that do not need to
change will not have changes (and will therefore also not be saved).