Allow calculators to implement adjustment labels by adjustable type
Prior to this, the `adjustment_label` method on
`SolidusPromotions::Benefit` could only output a relatively static
string. This commit allows calculators to define one of these methods:

- `#line_item_adjustment_label(line_item, options = {})`
- `#shipment_adjustment_label(shipment, options = {})`
- `#shipping_rate_adjustment_label(shipping_rate, options = {})`
- `#price_adjustment_label(price, options = {})`

These have access to the inner workings of the calculator, so they can
provide dynamic labels (to reflect e.g. which tier a promotion chooses,
making that transparent to the customer.